Dark Theme, Removed Fixed Navbar, and Minor Style Changes
Incorporated a dark theme and improved UI feedback for navbar and stream selections.  Removed the fixed navbar because when scrolling the stream container(s) partially show above the navbar and there is not an elegant way to fix this for chrome extensions.  Minor changes included removing a container that would create double scroll bars for the extension, increasing the height and width of the popup, and incorporating stream thumbnails into the stream selection(s).